в лучшем случае учитывая как криво прикручена ORM, это тянет на MVP. Постараюсь доделать на днях, а пока если будет время можешь посмотреть изменения.
