+++ ReportingSystem/Businesslayer/FormReport/IFormReport.cs
@@ -1,0 +1,9 @@
+﻿using ReportingSystem.DataAccesslayer.Report;
+
+namespace ReportingSystem.Businesslayer.FormReport
+{
+   public interface IFormReport
+   {
+       public Report GenerateReport();
+   }
+}
