+++ ReportingSystem/Businesslayer/TaskManagementSystem/Actions/IManageTask.cs
@@ -1,0 +1,7 @@
+﻿namespace ReportingSystem.Businesslayer.TaskManagementSystem.Actions
+{
+   public interface IManageTask
+   {
+        public DataAccesslayer.Task.Task SomeActionInTask();
+   }
+}
