+++ ReportingSystem/Businesslayer/TaskManagementSystem/SearchTask/IExecuteSearch.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace ReportingSystem.Businesslayer.TaskManagementSystem.SearchTask
+{
+    public interface IExecuteSearch
+    { 
+        public List<DataAccesslayer.Task.Task> ReturnTasks();
+    }
+}
