+++ ReportingSystem/Businesslayer/WorkersManagement/IManageWorkers.cs
@@ -1,0 +1,9 @@
+﻿using ReportingSystem.DataAccesslayer.Worker;
+
+namespace ReportingSystem.Businesslayer.WorkersManagement
+{
+    public interface IManageWorkers
+    {
+       public Worker Action();
+    }
+}
