--- ReportingSystem/Report/Actions/IFormReport.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ReportingSystem.Report.Actions
-{
-   public interface IFormReport
-   {
-       public Report GenerateReport();
-   }
-}
