--- ReportingSystem/Task/Actions/IManageTask.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ReportingSystem.Task
-{
-   public interface IManageTask
-   {
-        public Task SomeActionInTask();
-   }
-}
