--- ReportingSystem/Task/SearchTask/IExecuteSearch.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ReportingSystem.Task.SearchTask
-{
-    public interface IExecuteSearch
-    { 
-        public List<Task> ReturnTasks();
-    }
-}
