--- ReportingSystem/Worker/IManageWorkers.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ReportingSystem.Worker
-{
-    public interface IManageWorkers
-    {
-       public Worker Action();
-    }
-}
